+++ core/src/main/groovy/nz/govt/natlib/tools/sip/generation/fairfax/processor/FairfaxFilesProcessor.groovy
@@ -38,6 +38,8 @@ import java.time.LocalDate
  */
 @Log4j2
 class FairfaxFilesProcessor {
+    static final String FOREVER_PROJECT_PREFIX = "FP"
+
     FairfaxProcessingParameters processingParameters
     List<Path> filesForProcessing
 
@@ -114,6 +116,31 @@ class FairfaxFilesProcessor {
                 default:
                     sortedFilesForProcessing = []
                     break
+            }
+
+            // Determine if fairfaxFiles contains any files that starts with FP
+            // If so, process the files as a single collection with FP files at the end
+            boolean isFP = false
+            for (FairfaxFile file : sortedFilesForProcessing) {
+                if (file.getTitleCode().startsWith(FOREVER_PROJECT_PREFIX)) {
+                    isFP = true
+                    break
+                }
+            }
+            if (isFP) {
+                List<FairfaxFile> fpFiles = []
+                List<FairfaxFile> sortedFiles = []
+                for (FairfaxFile fairfaxFile : sortedFilesForProcessing) {
+                    if (fairfaxFile.getTitleCode().startsWith(FOREVER_PROJECT_PREFIX)) {
+                        fpFiles.add(fairfaxFile)
+                    } else {
+                        sortedFiles.add(fairfaxFile)
+                    }
+                }
+                for (FairfaxFile ff : fpFiles) {
+                    sortedFiles.add(ff)
+                }
+                sortedFilesForProcessing = sortedFiles
             }
 
             if (processingParameters.skip) {