+++ core/src/test/groovy/nz/govt/natlib/tools/sip/generation/fairfax/FairfaxSpreadsheetTest.groovy
@@ -1,6 +1,7 @@
 package nz.govt.natlib.tools.sip.generation.fairfax
 
 import static org.hamcrest.core.Is.is
+import static org.junit.Assert.assertFalse
 import static org.junit.Assert.assertThat
 import static org.junit.Assert.assertTrue
 
@@ -19,20 +20,20 @@ class FairfaxSpreadsheetTest {
         FairfaxSpreadsheet fairfaxSpreadsheet = FairfaxSpreadsheet.defaultInstance()
 
         assertTrue("Spreadsheet is valid", fairfaxSpreadsheet.spreadsheet.isValid(false, false))
-        List<Map<String, String>> mapsForCountryLivingList =
+        List<Map<String, String>> mapsForBayChronicleList =
                 fairfaxSpreadsheet.spreadsheet.mapsForColumn(FairfaxSpreadsheet.MMSID_COLUMN_NAME,
-                        "9918150268002836")
+                        "9917952983502836")
 
-        assertThat("Country Living only has one entry", mapsForCountryLivingList.size(), is(new Integer(1)))
-        Map<String, String> mapsForCountryLiving = mapsForCountryLivingList.first()
-        assertThat("'title_parent' is 'Country Living'", mapsForCountryLiving.get("title_parent"), is("Country Living"))
-        assertThat("'MMSID' is 9918150268002836", mapsForCountryLiving.get("MMSID"), is("9918150268002836"))
-        assertThat("'title_code' is 'CL4'", mapsForCountryLiving.get("title_code"), is("CL4"))
-        assertThat("'section_codes' is 'ED1'", mapsForCountryLiving.get("section_codes"), is("ED1"))
+        assertThat("The Bay Chronicle only has one entry", mapsForBayChronicleList.size(), is(new Integer(1)))
+        Map<String, String> mapsForBayChronicle = mapsForBayChronicleList.first()
+        assertThat("'title_parent' is 'The Bay Chronicle'", mapsForBayChronicle.get("title_parent"), is("The Bay Chronicle"))
+        assertThat("'MMSID' is 9917952983502836", mapsForBayChronicle.get("MMSID"), is("9917952983502836"))
+        assertThat("'title_code' is 'BAC'", mapsForBayChronicle.get("title_code"), is("BAC"))
+        assertThat("'section_codes' is 'ED1'", mapsForBayChronicle.get("section_codes"), is("ED1"))
 
-        assertThat("titleParent for titleCode: CL4 sectionCode: ED1 is 'Country Living'", fairfaxSpreadsheet.getTitleParentForTitleCodeSectionCode('CL4', 'ED1'),
-                is('Country Living'))
-        assertTrue("isMagazine is true for Country Living",
-                FairfaxSpreadsheet.extractBooleanValue(mapsForCountryLiving, FairfaxSpreadsheet.IS_MAGAZINE_KEY))
+        assertThat("titleParent for titleCode: BAC sectionCode: ED1 is 'The Bay Chronicle'", fairfaxSpreadsheet.getTitleParentForTitleCodeSectionCode('BAC', 'ED1'),
+                is('The Bay Chronicle'))
+        assertFalse("isMagazine is false forThe Bay Chronicle",
+                FairfaxSpreadsheet.extractBooleanValue(mapsForBayChronicle, FairfaxSpreadsheet.IS_MAGAZINE_KEY))
     }
 }