+++ docs/source/guides/release-notes.rst
@@ -6,7 +6,7 @@ Release Notes
 Introduction
 ============
 
-This guide, designed for a NLNZ Tools SIP Generation Fairfax developer, covers release notes from `1.0.0-SNAPSHOT`.
+This guide, designed for a NLNZ Tools SIP Generation Fairfax developer, covers release notes from `1.0.1-SNAPSHOT`.
 Versions are in reverse chronological order, with the most recent version first. While the
 *NLNZ Tools SIP Generation Fairfax Developer Guide* and *NLNZ Tools SIP Generation Fairfax User Guide* are accurate for
 the current release, the *Release Notes* can give some idea of how things have changed since the last major release.
@@ -20,7 +20,9 @@ Following this introduction, the Release Notes includes the following sections:
 
 -   **1.0.0** - Release 1.0.0.
 
--   **1.0.0-SNAPSHOT** - Current 1.0.0 development.
+-   **1.0.0-SNAPSHOT** - 1.0.0 development.
+
+-   **1.0.1-SNAPSHOT** - Current 1.0.1 development.
 
 
 Changes since 1.0.0
@@ -40,3 +42,9 @@ TODO Major features of this release
 ==============
 
 TODO Keep track of ongoing development changes here.
+
+
+1.0.1-SNAPSHOT
+==============
+
+The Forever Project includes files that are to be appended to the end of multiple titles.  As such the changes in 1.0.1-SNAPSHOT searches for Forever Project Files and appends them to the end of their respective titles to be processed.  In addition to this, sequence letter column was added to the spreadsheet and spreadsheet processor to handle supplement grouping files based on the use of sequence letters.